# docs/swagger.json
{"openapi":"3.0.1","info":{"title":"Game store API","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/v1/user":{"get":{"tags":["User"],"operationId":"get_1","parameters":[{"name":"userRetrievalRequestDto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/UserRetrievalRequestDTO"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRetrievalResponseDTO"}}}}}},"put":{"tags":["User"],"operationId":"update","requestBody":{"description":"UserUpdateRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}},"post":{"tags":["User"],"operationId":"create","requestBody":{"description":"UserCreationRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreationRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreationResponseDTO"}}}}}},"delete":{"tags":["User"],"operationId":"delete","requestBody":{"description":"UserDeleteRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDeleteRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/v1/game":{"get":{"tags":["Game"],"operationId":"get_7","parameters":[{"name":"gameRetrievalRequestDto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/GameRetrievalRequestDTO"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameRetrievalResponseDTO"}}}}}},"put":{"tags":["Game"],"operationId":"update_1","requestBody":{"description":"GameUpdateRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameUpdateRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}},"post":{"tags":["Game"],"operationId":"create_3","requestBody":{"description":"GameCreationRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameCreationRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameCreationResponseDTO"}}}}}},"delete":{"tags":["Game"],"operationId":"delete_3","requestBody":{"description":"GameDeleteRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameDeleteRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/v1/order":{"get":{"tags":["Order"],"operationId":"get_3","parameters":[{"name":"orderRetrievalRequestDTo","in":"query","required":true,"schema":{"$ref":"#/components/schemas/OrderRetrievalRequestDTO"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRetrievalResponseDTO"}}}}}},"post":{"tags":["Order"],"operationId":"create_1","requestBody":{"description":"OrderCreationRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCreationRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCreationResponseDTO"}}}}}},"delete":{"tags":["Order"],"operationId":"delete_1","requestBody":{"description":"OrderDeleteRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDeleteRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/v1/genre":{"get":{"tags":["Genre"],"operationId":"get_5","parameters":[{"name":"genreRetrievalRequestDto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/GenreRetrievalRequestDTO"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreRetrievalResponseDTO"}}}}}},"post":{"tags":["Genre"],"operationId":"create_2","requestBody":{"description":"GenreCreationRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreCreationRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreCreationResponseDTO"}}}}}},"delete":{"tags":["Genre"],"operationId":"delete_2","requestBody":{"description":"GenreDeleteRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreDeleteRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/v1/admin":{"get":{"tags":["Admin"],"operationId":"get_9","parameters":[{"name":"adminRetrievalRequestDto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/AdminRetrievalRequestDTO"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminRetrievalResponseDTO"}}}}}},"post":{"tags":["Admin"],"operationId":"create_4","requestBody":{"description":"AdminCreationRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminCreationRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminCreationResponseDTO"}}}}}},"delete":{"tags":["Admin"],"operationId":"delete_4","requestBody":{"description":"AdminDeleteRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminDeleteRequestDTO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK"}}}},"/v1/users":{"get":{"tags":["User"],"operationId":"get","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserRetrievalResponseDTO"}}}}}}}},"/v1/orders":{"get":{"tags":["Order"],"operationId":"get_2","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderRetrievalResponseDTO"}}}}}}}},"/v1/genres":{"get":{"tags":["Genre"],"operationId":"get_4","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GenreRetrievalResponseDTO"}}}}}}}},"/v1/games":{"get":{"tags":["Game"],"operationId":"get_6","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GameRetrievalResponseDTO"}}}}}}}},"/v1/admins":{"get":{"tags":["Admin"],"operationId":"get_8","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdminRetrievalResponseDTO"}}}}}}}}},"components":{"schemas":{"UserUpdateRequestDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"first_name":{"type":"string"},"last_name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"GameUpdateRequestDTO":{"type":"object","properties":{"name":{"type":"string"},"genres":{"type":"array","items":{"$ref":"#/components/schemas/Genre"}},"price":{"type":"integer","format":"int32"},"available_quantity":{"type":"integer","format":"int32"}}},"Genre":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"UserCreationRequestDTO":{"type":"object","properties":{"first_name":{"type":"string"},"last_name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"UserCreationResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"OrderCreationRequestDTO":{"type":"object","properties":{"cart_id":{"type":"integer","format":"int64"}}},"OrderCreationResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"GenreCreationRequestDTO":{"type":"object","properties":{"name":{"type":"string"}}},"GenreCreationResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"GameCreationRequestDTO":{"type":"object","properties":{"name":{"type":"string"},"genres":{"type":"array","items":{"$ref":"#/components/schemas/Genre"}},"price":{"type":"integer","format":"int32"},"available_quantity":{"type":"integer","format":"int32"}}},"GameCreationResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"AdminCreationRequestDTO":{"type":"object","properties":{"user_id":{"type":"integer","format":"int64"}}},"AdminCreationResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"UserRetrievalResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"first_name":{"type":"string"},"last_name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"UserRetrievalRequestDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"CartItemRetrievalResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserRetrievalResponseDTO"},"game":{"$ref":"#/components/schemas/GameRetrievalResponseDTO"},"quantity":{"type":"integer","format":"int32"}}},"CartRetrievalResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserRetrievalResponseDTO"},"cart_items":{"type":"array","items":{"$ref":"#/components/schemas/CartItemRetrievalResponseDTO"}}}},"GameRetrievalResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"genres":{"type":"array","items":{"$ref":"#/components/schemas/GenreRetrievalResponseDTO"}},"price":{"type":"integer","format":"int32"},"available_quantity":{"type":"integer","format":"int32"}}},"GenreRetrievalResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"OrderRetrievalResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"cart":{"$ref":"#/components/schemas/CartRetrievalResponseDTO"},"creation_time":{"type":"string","format":"date-time"}}},"OrderRetrievalRequestDTO":{"type":"object"},"GenreRetrievalRequestDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"GameRetrievalRequestDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"genres":{"type":"array","items":{"$ref":"#/components/schemas/Genre"}},"price_lt":{"type":"integer","format":"int32"},"price_gt":{"type":"integer","format":"int32"},"available_quantity":{"type":"integer","format":"int32"},"available_quantity_lt":{"type":"integer","format":"int32"},"available_quantity_gt":{"type":"integer","format":"int32"}}},"AdminRetrievalResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserRetrievalResponseDTO"}}},"AdminRetrievalRequestDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"UserDeleteRequestDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"OrderDeleteRequestDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"GenreDeleteRequestDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"GameDeleteRequestDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"AdminDeleteRequestDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}}}}}